Move non-API rewiring classes into impl package.
