+++ djvm/src/main/kotlin/net/corda/djvm/rewiring/impl/ThrowableWrapperFactory.kt
@@ -1,9 +1,10 @@
-package net.corda.djvm.rewiring
+package net.corda.djvm.rewiring.impl
 
 import net.corda.djvm.analysis.SyntheticResolver.Companion.isDJVMSynthetic
 import net.corda.djvm.code.CONSTRUCTOR_NAME
 import net.corda.djvm.code.DJVM_EXCEPTION_NAME
 import net.corda.djvm.code.DJVM_SYNTHETIC
+import net.corda.djvm.rewiring.ByteCode
 import org.objectweb.asm.ClassWriter
 import org.objectweb.asm.Opcodes.*
